requirements.txt
# ------------------ Backend ------------------
fastapi
uvicorn
pydantic
python-dotenv
requests

# ------------------ LLM / LangChain ------------------
langchain
langchain-core
langchain-community
langchain-openai
openai

# ------------------ Vector Store / Embeddings ------------------
faiss-cpu
sentence-transformers
torch

# ------------------ Git / Repo Handling ------------------
gitpython
tree-sitter

# ------------------ Frontend ------------------
streamlit

# ------------------ Database ------------------
supabase

# ------------------ Fernet Key ------------------
CREDENTIAL_ENCRYPTION_KEY=IVd2b5qHrgBsVOYkr1OkRs8HLnREoSIe1M5IGgbAS8Y=
